Remove BackPrompt in MassDrops.php

+++ modules/Scheduling/MassDrops.php
@@ -6,49 +6,68 @@ if(isset($_REQUEST['modfunc']) && $_REQUEST['modfunc']=='save')
 {
 	if($_SESSION['MassDrops.php'])
 	{
-		$END_DATE = $_REQUEST['day'].'-'.$_REQUEST['month'].'-'.$_REQUEST['year'];
-		if(!VerifyDate($END_DATE))
-			BackPrompt(_('The date you entered is not valid'));
-		$course_mp = DBGet(DBQuery("SELECT MARKING_PERIOD_ID FROM COURSE_PERIODS WHERE COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."'"));
-		$course_mp = $course_mp[1]['MARKING_PERIOD_ID'];
-		$course_mp_table = GetMP($course_mp,'MP');
-
-		if($course_mp_table!='FY' && $course_mp!=$_REQUEST['marking_period_id'] && mb_strpos(GetChildrenMP($course_mp_table,$course_mp),"'".$_REQUEST['marking_period_id']."'")===false)
-			BackPrompt(_('You cannot schedule a student into that course during this marking period.').' '.sprintf(_('This course meets on %s.'),GetMP($course_mp)));
-
-		$mp_table = GetMP($_REQUEST['marking_period_id'],'MP');
-		//$current_RET = DBGet(DBQuery("SELECT STUDENT_ID FROM SCHEDULE WHERE COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."' AND SYEAR='".UserSyear()."' AND (('".$start_date."' BETWEEN START_DATE AND END_DATE OR END_DATE IS NULL) AND '".$start_date."'>=START_DATE)"),array(),array('STUDENT_ID'));
-		$current_RET = DBGet(DBQuery("SELECT STUDENT_ID FROM SCHEDULE WHERE COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."' "));
-		foreach($_REQUEST['student'] as $student_id=>$yes)
+		if (is_array($_REQUEST['student']))
 		{
-			if(!$current_RET[$student_id])
+			$END_DATE = $_REQUEST['day'].'-'.$_REQUEST['month'].'-'.$_REQUEST['year'];
+			if(VerifyDate($END_DATE))
 			{
-				//echo "<b>student:</b>  $student_id ";
-				//$start_end_RET = DBGet(DBQuery("SELECT START_DATE,END_DATE FROM SCHEDULE WHERE STUDENT_ID='".UserStudentID()."' AND COURSE_PERIOD_ID='".$course_period_id."' AND END_DATE<START_DATE"));
-				//$start_end_RET = DBGet(DBQuery("SELECT START_DATE,END_DATE FROM SCHEDULE WHERE STUDENT_ID='".$student_id."' AND COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."' AND END_DATE<START_DATE"));
-				$start_end_RET = DBGet(DBQuery("SELECT START_DATE,END_DATE FROM SCHEDULE WHERE STUDENT_ID='".$student_id."' AND COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."'"));
-				if(count($start_end_RET))
+				$course_mp = DBGet(DBQuery("SELECT MARKING_PERIOD_ID FROM COURSE_PERIODS WHERE COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."'"));
+				$course_mp = $course_mp[1]['MARKING_PERIOD_ID'];
+				$course_mp_table = GetMP($course_mp,'MP');
+
+				if($course_mp_table!='FY' && $course_mp!=$_REQUEST['marking_period_id'] && mb_strpos(GetChildrenMP($course_mp_table,$course_mp),"'".$_REQUEST['marking_period_id']."'")===false)
+					BackPrompt(_('You cannot schedule a student into that course during this marking period.').' '.sprintf(_('This course meets on %s.'),GetMP($course_mp)));
+
+				$mp_table = GetMP($_REQUEST['marking_period_id'],'MP');
+				//$current_RET = DBGet(DBQuery("SELECT STUDENT_ID FROM SCHEDULE WHERE COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."' AND SYEAR='".UserSyear()."' AND (('".$start_date."' BETWEEN START_DATE AND END_DATE OR END_DATE IS NULL) AND '".$start_date."'>=START_DATE)"),array(),array('STUDENT_ID'));
+				$current_RET = DBGet(DBQuery("SELECT STUDENT_ID FROM SCHEDULE WHERE COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."' "));
+				foreach($_REQUEST['student'] as $student_id=>$yes)
 				{
-					//DBQuery("DELETE FROM SCHEDULE WHERE STUDENT_ID='".UserStudentID()."' AND END_DATE IS NOT NULL AND END_DATE<START_DATE");
-					DBQuery("DELETE FROM SCHEDULE WHERE STUDENT_ID='".$student_id."' AND COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."'");
-					
-//modif Francois: Moodle integrator
-					$moodleError .= Moodle($_REQUEST['modname'], 'core_role_unassign_roles');
+					if(!$current_RET[$student_id])
+					{
+						//echo "<b>student:</b>  $student_id ";
+						//$start_end_RET = DBGet(DBQuery("SELECT START_DATE,END_DATE FROM SCHEDULE WHERE STUDENT_ID='".UserStudentID()."' AND COURSE_PERIOD_ID='".$course_period_id."' AND END_DATE<START_DATE"));
+						//$start_end_RET = DBGet(DBQuery("SELECT START_DATE,END_DATE FROM SCHEDULE WHERE STUDENT_ID='".$student_id."' AND COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."' AND END_DATE<START_DATE"));
+						$start_end_RET = DBGet(DBQuery("SELECT START_DATE,END_DATE FROM SCHEDULE WHERE STUDENT_ID='".$student_id."' AND COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."'"));
+						if(count($start_end_RET))
+						{
+							//DBQuery("DELETE FROM SCHEDULE WHERE STUDENT_ID='".UserStudentID()."' AND END_DATE IS NOT NULL AND END_DATE<START_DATE");
+							DBQuery("DELETE FROM SCHEDULE WHERE STUDENT_ID='".$student_id."' AND COURSE_PERIOD_ID='".$_SESSION['MassDrops.php']['course_period_id']."'");
+							
+		//modif Francois: Moodle integrator
+							$moodleError .= Moodle($_REQUEST['modname'], 'core_role_unassign_roles');
+						}
+					}
 				}
+				$note[] = '<IMG SRC="assets/check.png" class="alignImg">&nbsp;'._('This course has been dropped for the selected students\' schedules.');
 			}
+			else
+				$error[] = _('The date you entered is not valid');
 		}
-		unset($_REQUEST['modfunc']);
-		unset($_SESSION['MassDrops.php']);
-		$note = _('This course has been dropped for the selected students\' schedules.');
+		else
+			$error[] = _('You must choose at least one student.');
 	}
 	else
-		BackPrompt(_('You must choose a course.'));
+		$error[] = _('You must choose a course.');
+		
+	unset($_SESSION['_REQUEST_vars']['modfunc']);
+	unset($_REQUEST['modfunc']);
+	unset($_SESSION['MassDrops.php']);
 }
 
 
 if($_REQUEST['modfunc']!='choose_course')
 {
 	DrawHeader(ProgramTitle());
+	
+	if (isset($error))
+		echo ErrorMessage($error);
+	if(isset($note))
+		echo ErrorMessage($note, 'note');
+
+//modif Francois: Moodle integrator
+	echo $moodleError;
+
 	if($_REQUEST['search_modfunc']=='list')
 	{
 		echo '<FORM action="Modules.php?modname='.$_REQUEST['modname'].'&modfunc=save" method="POST">';
@@ -79,12 +98,6 @@ if($_REQUEST['modfunc']!='choose_course')
 		echo '</TABLE></TD></TR></TABLE><BR />';
 	}
 
-	if($note)
-//		DrawHeader('<IMG SRC=assets/check.png>'.$note);
-		echo '<div class="updated"><IMG SRC="assets/check.png" class="alignImg">&nbsp;'.$note.'</div>';
-
-//modif Francois: Moodle integrator
-	echo $moodleError;
 }
 
 if(empty($_REQUEST['modfunc']))
